Fixed weird bugs with showMenu(), and a SIGBUS error in apps_opt_cmdline().

+++ lib/menu.h
@@ -19,6 +19,6 @@
 
 #include <stddef.h>
 
-int showMenu(char *prefix, char *title, char **options, size_t numOptions);
+int showMenu(char *prefix, char *title, char **options, int numOptions);
 
 #endif
